+++ query-node/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Refactor of mappings for more better handling of error cases. [#4856](https://github.com/Joystream/joystream/pull/4856)
 - Bug fix [#4855](https://github.com/Joystream/joystream/issues/4855)
-- Add support for UpdateGlobalNftLimit proposal. 
+- Add support for UpdateGlobalNftLimit proposal.
 
 ### 1.6.0